# openapi/v2.json
{
  "components": {
...
    "schemas": {
...
      "EncourageRequest": {
        "properties": {
          "message_key": {
            "example": "great_work",
            "type": "string"
          },
          "student_id": {
            "example": "550e8400-e29b-41d4-a716-446655440000",
            "format": "uuid",
            "type": "string"
          }
        },
        "required": [
          "student_id"
        ],
        "type": "object"
      },
      "ErrorResponse": {
        "properties": {
          "code": {
            "example": "VALIDATION_ERROR",
            "type": "string"
          },
          "error": {
            "example": "Validation failed",
            "type": "string"
          },
          "success": {
            "enum": [
              false
            ],
            "type": "boolean"
          }
        },
        "required": [
          "success",
          "error"
        ],
        "type": "object"
      },
...
      "SuccessAck": {
        "properties": {
          "success": {
            "enum": [
              true
            ],
            "type": "boolean"
          }
        },
        "required": [
          "success"
        ],
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "bearerAuth": {
        "bearerFormat": "JWT",
        "description": "Supabase access-token JWT (mobile clients).",
        "scheme": "bearer",
        "type": "http"
      },
      "cookieAuth": {
        "description": "Supabase session cookie (web clients).",
        "in": "cookie",
        "name": "sb-access-token",
        "type": "apiKey"
      }
    }
  },
  "info": {
    "description": "The Alfanumrik /v2 contract. Single source of truth: src/lib/api/v2/contract.ts (Zod). This document is generated by `npm run gen:openapi` and consumed by the Dart client codegen (`npm run gen:dart`). Do not edit by hand.",
    "title": "Alfanumrik /v2 API",
    "version": "2.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/v2/parent/encourage": {
      "post": {
        "description": "Parent sends a curated, preset-keyed encouragement to a linked child. Requires child.encourage and an approved guardian↔student link. Rate-limited to one cheer per (guardian, student) per 6 hours.",
        "operationId": "postParentEncourage",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EncourageRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessAck"
                }
              }
            },
            "description": "Cheer sent."
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Invalid body (bad student_id or unknown message_key)."
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "No parent profile, or not linked to this student."
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Rate limited — already cheered within the last 6 hours."
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unexpected server error."
          },
          "502": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Notification fan-out failed."
          }
        },
        "security": [
          {
            "bearerAuth": []
          },
          {
            "cookieAuth": []
          }
        ],
        "summary": "Send a preset cheer to a linked child",
        "tags": [
          "parent"
        ]
      }
    },
...
  },
  "servers": [
    {
      "description": "Alfanumrik API base path",
      "url": "/api"
    }
  ],
...
}